CB-14926 [e2e] WaitObject: introduce ignoreFailedState -- phase II. of CB-14893
In CB-14893 the await method was further customized: it not only waits for an object to reach a status, but the test writer can specify to ignore some of the failure states, called ignoredFailedStatuses. CB-14893 gives explanation on the why. Also, that commit gave full implementation for SDXWaitObject only.

This commit extends ignoredFailedStatuses to all other WaitObjects. It is still used only in some SDX tests.